chore: update tests for grading selectors
